// src/app/models/funcionario.model.ts
export interface Funcionario {
  id: number;
  usuario: string;
  senha: number;
  nome: string;
  idade: number;
  sexo: string;
  cpf: string;
  rua: string;
  numero: string;
  complemento: string;
  bairro: string;
  cidade: string;
  estado: string;
  contato: string;
  email: string;
  dataNascimento: string;
}